Creating orders
I have added the functionality of creating new orders using react-router's custom form components as well as a few of its hooks.

One thing i have neglected in my earlier projects is error handling and so rather than add it in at the end i feel i will this time add them as i go along. So i have handled all the errors related to the getting or posting of data to and from the custom API i am using for this.