Tailwind
I have begun learning tailwind CSS as i have encountered a few job posts that are looking for experience with it. So i spent my time refactoring and styling today along side a ton of reading up and learning about tailwind. Its in fact quite easy with the help of the docs to memorize the changed property names it seems mostly if you know the css properties you can find anything you need to know so i feel i will be mostly just getting used to using it and trying to memorize the necessary class names. Tailwind also seems to favour mobile first development and so i am adapting my approach to suit that. So far i feel it could be an incredibly useful tool in speeding up my workflow despite the face it does clutter the JSX / HTML classes.