Styling Components with tailwind
I have been a bit neglectful of my commits over the last few days and so they are all bundled into this one, i have used what i have learned so far to style most of the remaining components to my liking using all the things ive learned along the way.

I have also had a play around with tailwinds configuration options and can see the power and freedom to customize tailwind that comes along with it. Im very excited to continue using it for now. As a point of personal preference however i am not a huge fan of writing the styles into the class names. I do much prefer my CSS to be abstracted to a new file so i will continue that route for now. I felt tailwind would be useful in landing a job that may operate with a mobile first development strategy.