Introduction of Redux
Today i have gotten started laying the ground work for redux and redux toolkit. I have created a store and a userSlice to store all the global state related to the user. Right now this is only the username but it will be much more later, i have also connected it up to my Username component to display the name from the state and altered main.jsx to next the App component in the redu Provider component.