Redux Cart implementation
I have implemented Redux to the Cart component to control the actual cart as we are displaying this cart not only on that page but on others in the application as well such as the menu page. So i have added a cart slice and given it the necessary reducers for adding, deleting, clearing the cart and controlling the quantity of the cart items. Although the quantity feature is not fully functional yet. We also made a new DeleteItem component because again this will be used on both the cart itself and the menu pages and so it was necessary to abstract this to its own component.