Item quantities and placing new orders
I have added the functionality for increasing or decreasing the quantity of an item by adding the UpdateItemQuantity component. I have also added the logic for removing the item from the cart if the user tries to push the quantity lower than 1. If so the item will be removed from the cart and the price unchanged. I have also added the functionality of placing the order and also the logic for the priority cost. Then i amended the UI to better display to the user all the data they need.