Redux thunks and Geolocation
Today i have added the functionality for the geolocation services to autofill the address with the users permission when submitting an order. To be able to do this i needed async behaviour and redux is synchronous, im still somewhat unfamiliar with thunks but will change that very shortly. From what i understand they allow us to perform some behavior after dispatching an action but before it reaches the reducer allowing us to perform some asynchronous behavior on it in this case. We can then write some reducer functions to handle the 3 potential cases we get back from our thunk, being the pending, fulfilled and rejected cases to achieve our goal of setting the address and position for the user.